Piece/board structure overhaul
Added piece class, and a subclass for each piece with a method to get the piece's possible moves.

Restructured script to use a 2D array as the chess board.
The 2D array contains either an instance of a piece subclass or null. The piece instances contain relevant information about the piece in that slot.

Temporary functionalty: clicking each piece console.logs an array of objects containing every possible x and y coordinate that piece can move to from its current position. Does not account for collision with other pieces yet.