Pawn logic + Temp feature
Basic logic for getting pawns is functional, but definitely not optimal yet. It does not account for en passant yet, as we don't have a way to move pieces at all yet.

Also refactored big parts of the pieces' getPossibleMoves method to use a helper method in the parent Piece class to prevent repeating similar calculations in code

Temp feature for testing collision of backline pieces: Alt + click to delete a piece on the board!