Pawn refactor + show possible moves
Refactored the logic for getting possible pawn moves, still a bit clunky but a lot better than it was before.

Changed the for loop from < to <=. Corrected all max moves accordingly (because a piece will never move 8 tiles, 7 tiles is the max if you're on the edge of the board, because the piece itself takes up 1 tile.)

Now uses the returned object from getPossibleMoves to visualize the tiles you can interact with. Does not include logic for actually moving any pieces yet.